format: fix -Wconversion in j2t_indent (size_t indent counter)
j2t_indent computed the indent width as `unsigned n` but then did
`n -= sizeof spaces`, mixing unsigned with size_t; under -Werror=conversion
(the `make strict` / CI strict lane on gcc) narrowing the result back to
unsigned is an error. Widen n to size_t and compute (size_t)level *
indent_width so the multiply cannot overflow unsigned either. No behavior
change.

+++ src/format.c
@@ -55,7 +55,7 @@ void j2t_putc(j2t_out *o, char c) {
 void j2t_puts(j2t_out *o, const char *s) { j2t_write(o, s, strlen(s)); }
 
 void j2t_indent(j2t_out *o, unsigned level) {
-  unsigned n = level * o->indent_width;
+  size_t n = (size_t)level * o->indent_width;
   static const char spaces[16] =
       {' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' ',' '};
   while (n >= sizeof spaces) {
